refactor: you don't always have to inject dependencies
